# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
      "AcrossDatabases" : {
        "properties" : {
          "confirm" : {
            "$ref" : "#/components/schemas/Confirm"
          },
          "recommended" : {
            "items" : {
              "$ref" : "#/components/schemas/IndexRecommendation"
            },
            "type" : "array"
          }
        },
        "required" : [ "confirm", "recommended" ],
        "type" : "object"
      },
...
      "Applied" : {
        "properties" : {
          "caveat" : {
            "type" : "string"
          },
          "database" : {
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/AppliedIndex"
            },
            "type" : "array"
          }
        },
        "required" : [ "caveat", "database", "items" ],
        "type" : "object"
      },
      "AppliedIndex" : {
        "properties" : {
          "appliedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "database" : {
            "type" : "string"
          },
          "ddl" : {
            "type" : "string"
          },
          "index" : {
            "type" : "string"
          },
          "queries" : {
            "items" : {
              "$ref" : "#/components/schemas/AppliedQuery"
            },
            "type" : "array"
          },
          "table" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "appliedAt", "database", "ddl", "index", "queries", "table" ],
        "type" : "object"
      },
      "AppliedQuery" : {
        "properties" : {
          "callsAfter" : {
            "format" : "int64",
            "type" : "integer"
          },
          "callsBefore" : {
            "format" : "int64",
            "type" : "integer"
          },
          "measuredChangeFraction" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "measuredMeanMsAfter" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "measuredMeanMsBefore" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "plannerCostDropFraction" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "queryid" : {
            "type" : "string"
          },
          "sqlPreview" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "callsAfter", "callsBefore", "measuredChangeFraction", "measuredMeanMsAfter", "measuredMeanMsBefore", "plannerCostDropFraction", "queryid", "sqlPreview", "status" ],
        "type" : "object"
      },
      "Confirm" : {
        "properties" : {
          "caveat" : {
            "type" : "string"
          },
          "command" : {
            "type" : "string"
          }
        },
        "required" : [ "caveat", "command" ],
        "type" : "object"
      },
...
      "DatabaseSummary" : {
        "properties" : {
          "agent" : {
            "enum" : [ "NEVER_CONNECTED", "CONNECTED", "STALE" ],
            "type" : "string"
          },
          "appliedIndexes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "calls" : {
            "format" : "int64",
            "type" : "integer"
          },
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : "number"
          },
          "indexesToConsider" : {
            "format" : "int32",
            "type" : "integer"
          },
          "lastSampleAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "measuredMsWithAdvice" : {
            "format" : "double",
            "type" : "number"
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "name" : {
            "type" : "string"
          },
          "queries" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "agent", "appliedIndexes", "calls", "estimatedMsSaved", "indexesToConsider", "lastSampleAt", "measuredMsWithAdvice", "measuredTotalMs", "name", "queries" ],
        "type" : "object"
      },
      "DbTimeline" : {
        "properties" : {
          "database" : {
            "type" : "string"
          },
          "from" : {
            "format" : "date-time",
            "type" : "string"
          },
          "points" : {
            "items" : {
              "$ref" : "#/components/schemas/DbTimelinePoint"
            },
            "type" : "array"
          },
          "series" : {
            "items" : {
              "$ref" : "#/components/schemas/Series"
            },
            "type" : "array"
          },
          "to" : {
            "format" : "date-time",
            "type" : "string"
          },
          "window" : {
            "type" : "string"
          }
        },
        "required" : [ "database", "from", "points", "series", "to", "window" ],
        "type" : "object"
      },
      "DbTimelinePoint" : {
        "properties" : {
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : "number"
          },
          "hour" : {
            "format" : "date-time",
            "type" : "string"
          },
          "measuredMsBySeries" : {
            "items" : {
              "format" : "double",
              "type" : "number"
            },
            "type" : "array"
          },
          "measuredMsWithAdvice" : {
            "format" : "double",
            "type" : "number"
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "sampled" : {
            "type" : "boolean"
          }
        },
        "required" : [ "estimatedMsSaved", "hour", "measuredMsBySeries", "measuredMsWithAdvice", "measuredTotalMs", "sampled" ],
        "type" : "object"
      },
...
      "Hygiene" : {
        "properties" : {
          "database" : {
            "type" : "string"
          },
          "findings" : {
            "items" : {
              "$ref" : "#/components/schemas/IndexHygieneFinding"
            },
            "type" : "array"
          }
        },
        "required" : [ "database", "findings" ],
        "type" : "object"
      },
      "IndexHygieneFinding" : {
        "properties" : {
          "definition" : {
            "type" : "string"
          },
          "indexName" : {
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "DUPLICATE", "REDUNDANT", "UNUSED" ],
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "relatedIndex" : {
            "type" : [ "string", "null" ]
          },
          "table" : {
            "type" : "string"
          }
        },
        "required" : [ "definition", "indexName", "kind", "reason", "relatedIndex", "table" ],
        "type" : "object"
      },
      "IndexRecommendation" : {
        "properties" : {
          "accessMethod" : {
            "type" : "string"
          },
          "actionable" : {
            "type" : "boolean"
          },
          "buildCaution" : {
            "type" : [ "string", "null" ]
          },
          "database" : {
            "type" : "string"
          },
          "ddl" : {
            "type" : "string"
          },
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : "number"
          },
          "footprintLabel" : {
            "type" : [ "string", "null" ]
          },
          "queries" : {
            "items" : {
              "$ref" : "#/components/schemas/QueryEvidence"
            },
            "type" : "array"
          },
          "redundantWith" : {
            "type" : [ "string", "null" ]
          },
          "table" : {
            "type" : [ "string", "null" ]
          },
          "writeLoad" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/TableWriteLoad"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "accessMethod", "actionable", "buildCaution", "database", "ddl", "estimatedMsSaved", "footprintLabel", "queries", "redundantWith", "table", "writeLoad" ],
        "type" : "object"
      },
...
      "NewSlow" : {
        "properties" : {
          "database" : {
            "type" : "string"
          },
          "from" : {
            "format" : "date-time",
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/NewSlowView"
            },
            "type" : "array"
          },
          "minTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "to" : {
            "format" : "date-time",
            "type" : "string"
          },
          "window" : {
            "type" : "string"
          }
        },
        "required" : [ "database", "from", "items", "minTotalMs", "to", "window" ],
        "type" : "object"
      },
      "NewSlowView" : {
        "properties" : {
          "calls" : {
            "format" : "int64",
            "type" : "integer"
          },
          "firstSeen" : {
            "format" : "date-time",
            "type" : "string"
          },
          "measuredMeanMs" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "queryid" : {
            "type" : "string"
          },
          "sql" : {
            "type" : "string"
          }
        },
        "required" : [ "calls", "firstSeen", "measuredMeanMs", "measuredTotalMs", "queryid", "sql" ],
        "type" : "object"
      },
...
      "NotPlannerValidated" : {
        "properties" : {
          "accessMethod" : {
            "type" : "string"
          },
          "ddl" : {
            "type" : "string"
          },
          "queryids" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "reason" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "accessMethod", "ddl", "queryids", "reason" ],
        "type" : "object"
      },
      "Overview" : {
        "properties" : {
          "applied" : {
            "items" : {
              "$ref" : "#/components/schemas/AppliedIndex"
            },
            "type" : "array"
          },
          "appliedCaveat" : {
            "type" : "string"
          },
          "calls" : {
            "format" : "int64",
            "type" : "integer"
          },
          "confirm" : {
            "$ref" : "#/components/schemas/Confirm"
          },
          "databases" : {
            "items" : {
              "$ref" : "#/components/schemas/DatabaseSummary"
            },
            "type" : "array"
          },
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : "number"
          },
          "from" : {
            "format" : "date-time",
            "type" : "string"
          },
          "indexesToConsider" : {
            "format" : "int32",
            "type" : "integer"
          },
          "measuredMsWithAdvice" : {
            "format" : "double",
            "type" : "number"
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "notPlannerValidated" : {
            "format" : "int32",
            "type" : "integer"
          },
          "queries" : {
            "format" : "int64",
            "type" : "integer"
          },
          "timeline" : {
            "items" : {
              "$ref" : "#/components/schemas/TimelinePoint"
            },
            "type" : "array"
          },
          "to" : {
            "format" : "date-time",
            "type" : "string"
          },
          "topRecommendations" : {
            "items" : {
              "$ref" : "#/components/schemas/IndexRecommendation"
            },
            "type" : "array"
          },
          "window" : {
            "type" : "string"
          }
        },
        "required" : [ "applied", "appliedCaveat", "calls", "confirm", "databases", "estimatedMsSaved", "from", "indexesToConsider", "measuredMsWithAdvice", "measuredTotalMs", "notPlannerValidated", "queries", "timeline", "to", "topRecommendations", "window" ],
        "type" : "object"
      },
...
      "QueryEvidence" : {
        "properties" : {
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : "number"
          },
          "plannerCostDropFraction" : {
            "format" : "double",
            "type" : "number"
          },
          "queryid" : {
            "type" : "string"
          },
          "rangeLabel" : {
            "type" : [ "string", "null" ]
          },
          "scoreBasis" : {
            "type" : [ "string", "null" ]
          },
          "sqlPreview" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "estimatedMsSaved", "plannerCostDropFraction", "queryid", "rangeLabel", "scoreBasis", "sqlPreview" ],
        "type" : "object"
      },
...
      "Recommendations" : {
        "properties" : {
          "confirm" : {
            "$ref" : "#/components/schemas/Confirm"
          },
          "database" : {
            "type" : "string"
          },
          "notPlannerValidated" : {
            "items" : {
              "$ref" : "#/components/schemas/NotPlannerValidated"
            },
            "type" : "array"
          },
          "recommended" : {
            "items" : {
              "$ref" : "#/components/schemas/IndexRecommendation"
            },
            "type" : "array"
          }
        },
        "required" : [ "confirm", "database", "notPlannerValidated", "recommended" ],
        "type" : "object"
      },
      "Series" : {
        "properties" : {
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "queryid" : {
            "type" : "string"
          },
          "sqlPreview" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "estimatedMsSaved", "measuredTotalMs", "queryid", "sqlPreview" ],
        "type" : "object"
      },
...
      "TableActivity" : {
        "properties" : {
          "deleted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "inserted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tuplesRead" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updated" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "deleted", "inserted", "tuplesRead", "updated" ],
        "type" : "object"
      },
      "TableWriteLoad" : {
        "properties" : {
          "activity" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/TableActivity"
            }, {
              "type" : "null"
            } ]
          },
          "label" : {
            "type" : "string"
          },
          "level" : {
            "enum" : [ "WRITE_DOMINANT", "READ_DOMINANT", "NO_ACTIVITY" ],
            "type" : "string"
          },
          "table" : {
            "type" : "string"
          },
          "window" : {
            "type" : "string"
          }
        },
        "required" : [ "activity", "label", "level", "table", "window" ],
        "type" : "object"
      },
      "TimelinePoint" : {
        "properties" : {
          "estimatedMsSaved" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "hour" : {
            "format" : "date-time",
            "type" : "string"
          },
          "measuredMsWithAdvice" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "measuredTotalMs" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          }
        },
        "required" : [ "estimatedMsSaved", "hour", "measuredMsWithAdvice", "measuredTotalMs" ],
        "type" : "object"
      },
      "TopMoverView" : {
        "properties" : {
          "deltaMs" : {
            "format" : "double",
            "type" : "number"
          },
          "measuredPriorTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "measuredRecentTotalMs" : {
            "format" : "double",
            "type" : "number"
          },
          "pctChange" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "queryid" : {
            "type" : "string"
          },
          "sql" : {
            "type" : "string"
          }
        },
        "required" : [ "deltaMs", "measuredPriorTotalMs", "measuredRecentTotalMs", "pctChange", "queryid", "sql" ],
        "type" : "object"
      },
      "TopMovers" : {
        "properties" : {
          "database" : {
            "type" : "string"
          },
          "from" : {
            "format" : "date-time",
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/TopMoverView"
            },
            "type" : "array"
          },
          "to" : {
            "format" : "date-time",
            "type" : "string"
          },
          "window" : {
            "type" : "string"
          }
        },
        "required" : [ "database", "from", "items", "to", "window" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "description" : "A session token (login) or an API token (read-only).",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Read the slow-query history, plans, index recommendations and explanations PgLens keeps for each monitored database. queryid is a string everywhere (a signed 64-bit value). Numbers named measured… are summed pg_stat_statements deltas; estimated… and plannerCost… are planner estimates.",
    "license" : {
      "name" : "Apache-2.0"
    },
    "title" : "PgLens API",
    "version" : "0.1.0-rc"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/databases/{db}/applied" : {
      "get" : {
        "operationId" : "applied",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Applied"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "impact-controller" ]
      }
    },
    "/api/v1/databases/{db}/hygiene" : {
      "get" : {
        "operationId" : "hygiene",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Hygiene"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insight-controller" ]
      }
    },
    "/api/v1/databases/{db}/new-slow" : {
      "get" : {
        "operationId" : "newSlow",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "window",
          "required" : false,
          "schema" : {
            "default" : "24h",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "minTotalMs",
          "required" : false,
          "schema" : {
            "default" : 1000.0,
            "format" : "double",
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NewSlow"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insight-controller" ]
      }
    },
...
    "/api/v1/databases/{db}/recommendations" : {
      "get" : {
        "operationId" : "recommendations",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Recommendations"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insight-controller" ]
      }
    },
    "/api/v1/databases/{db}/timeline" : {
      "get" : {
        "operationId" : "timeline",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "window",
          "required" : false,
          "schema" : {
            "default" : "24h",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "series",
          "required" : false,
          "schema" : {
            "default" : 5,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DbTimeline"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "impact-controller" ]
      }
    },
    "/api/v1/databases/{db}/top-movers" : {
      "get" : {
        "operationId" : "topMovers",
        "parameters" : [ {
          "in" : "path",
          "name" : "db",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "window",
          "required" : false,
          "schema" : {
            "default" : "7d",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopMovers"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insight-controller" ]
      }
    },
...
    "/api/v1/overview" : {
      "get" : {
        "operationId" : "overview",
        "parameters" : [ {
          "in" : "query",
          "name" : "window",
          "required" : false,
          "schema" : {
            "default" : "24h",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Overview"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "impact-controller" ]
      }
    },
    "/api/v1/recommendations" : {
      "get" : {
        "operationId" : "acrossDatabases",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AcrossDatabases"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insight-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}